Feedback fixes

+++ database/migrations/2020_11_04_192014_create_invoices_table.php
@@ -12,11 +12,11 @@ class CreateInvoicesTable extends Migration
         Schema::create('invoices', function (Blueprint $table) {
             $table->id();
             $table->enum('type_invoice', ['cliente', 'remitente']);
-            $table->integer('commission')->default(0);
+            $table->float('commission')->default(0);
             $table->integer('commission_percentage');
             $table->float('partial_payment')->default(0);
             $table->float('subtotal')->default(0);
-            $table->float('total')->default(0);        
+            $table->float('total')->default(0);
             $table->boolean('is_price_modified')->default(false);
             $table->float('price_modified')->default(0);
             $table->text('modified_description')->default('');
